backend-fastapi/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.0
httpx==0.27.2
python-jose[cryptography]==3.3.0
pydantic==2.9.2
orjson==3.10.7
redis==5.2.1
pynacl==1.5.0
